Had a quick go at creating a basic UI with no game logic.
Implemented a UI with no game logic, based on my initial implementation
of the game. The UI supports resizing and setting the board to an
arbitrary size.

Tried to keep the code super simple and commented enough so that dojo
attendees can quickly figure out where to drop in the calls to their
game engine.

Still needs some work to do some UI checking for available flags and
also notifying users when the game is complete or lost.